home-automation/simulator/requirements.txt
azure-core==1.30.2
azure-cosmos==4.7.0
azure-eventgrid==4.21.0
azure-identity==1.17.1
azure-storage-blob==12.22.0
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.8
cryptography==43.0.0
Flask==3.1.0
idna==3.8
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
msal==1.30.0
msal-extensions==1.2.0
ngrok==1.4.0
portalocker==2.10.1
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.1.3
